% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{.optidesign_z}
\alias{.optidesign_z}
\title{Internal z-score helper}
\usage{
.optidesign_z(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
Internal z-score helper
}
\keyword{internal}
