% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{.optidesign_dbg}
\alias{.optidesign_dbg}
\title{Internal debug message helper}
\usage{
.optidesign_dbg(debug, ...)
}
\arguments{
\item{debug}{Logical.}

\item{...}{Arguments passed to \code{sprintf()}.}
}
\description{
Internal debug message helper
}
\keyword{internal}
